refactor code into Utilities project
