+++ NzWalksAPI.sln
@@ -3,7 +3,17 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NzWalksAPI", "NzWalksAPI.csproj", "{3AD55549-49B2-487C-A8BF-FC044453876F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NzWalksAPI.API", "NzWalksAPI.API\NzWalksAPI.API.csproj", "{94BBB82F-417E-4E72-8E46-2FF28FE43C7A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NzWalksAPI.Data", "NzWalksAPI.Data\NzWalksAPI.Data.csproj", "{B771FD02-7DFC-4DC2-961F-F6DBB8822098}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NzWalksAPI.Domain", "NzWalksAPI.Domain\NzWalksAPI.Domain.csproj", "{9C7C4F63-5FE1-42D1-83A2-107FFF45D878}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NzWalksAPI.Utilities", "NzWalksAPI.Utilities\NzWalksAPI.Utilities.csproj", "{15057C56-891F-4368-A9B6-CFAF94FDB099}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NzWalksAPI.Mapping", "NzWalksAPI.Mapping\NzWalksAPI.Mapping.csproj", "{DC166948-1B09-412C-9F58-C8A7D58D6834}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NzWalksAPI.Core", "NzWalksAPI.Core\NzWalksAPI.Core.csproj", "{7F5DB9D7-DD20-4995-A856-96502496263A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +21,30 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3AD55549-49B2-487C-A8BF-FC044453876F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3AD55549-49B2-487C-A8BF-FC044453876F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3AD55549-49B2-487C-A8BF-FC044453876F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3AD55549-49B2-487C-A8BF-FC044453876F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{94BBB82F-417E-4E72-8E46-2FF28FE43C7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{94BBB82F-417E-4E72-8E46-2FF28FE43C7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{94BBB82F-417E-4E72-8E46-2FF28FE43C7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{94BBB82F-417E-4E72-8E46-2FF28FE43C7A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B771FD02-7DFC-4DC2-961F-F6DBB8822098}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B771FD02-7DFC-4DC2-961F-F6DBB8822098}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B771FD02-7DFC-4DC2-961F-F6DBB8822098}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B771FD02-7DFC-4DC2-961F-F6DBB8822098}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9C7C4F63-5FE1-42D1-83A2-107FFF45D878}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9C7C4F63-5FE1-42D1-83A2-107FFF45D878}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9C7C4F63-5FE1-42D1-83A2-107FFF45D878}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9C7C4F63-5FE1-42D1-83A2-107FFF45D878}.Release|Any CPU.Build.0 = Release|Any CPU
+		{15057C56-891F-4368-A9B6-CFAF94FDB099}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{15057C56-891F-4368-A9B6-CFAF94FDB099}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{15057C56-891F-4368-A9B6-CFAF94FDB099}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{15057C56-891F-4368-A9B6-CFAF94FDB099}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DC166948-1B09-412C-9F58-C8A7D58D6834}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DC166948-1B09-412C-9F58-C8A7D58D6834}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DC166948-1B09-412C-9F58-C8A7D58D6834}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DC166948-1B09-412C-9F58-C8A7D58D6834}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7F5DB9D7-DD20-4995-A856-96502496263A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7F5DB9D7-DD20-4995-A856-96502496263A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7F5DB9D7-DD20-4995-A856-96502496263A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7F5DB9D7-DD20-4995-A856-96502496263A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
